add robobrain2 inference

+++ VLN_CE/habitat_extensions/actions.py
@@ -1,5 +1,5 @@
 from typing import Any
-
+# ! change np.float -> np.float32
 import numpy as np
 from gym import spaces
 from habitat.core.registry import registry
@@ -63,12 +63,12 @@ class GoTowardPoint(TeleportAction):
                 "r": spaces.Box(
                     low=np.array([0.0]),
                     high=np.array([np.sqrt(2 * (coord_range ** 2))]),
-                    dtype=np.float,
+                    dtype=np.float32,
                 ),
                 "theta": spaces.Box(
                     low=np.array([0.0]),
                     high=np.array([2 * np.pi]),
-                    dtype=np.float,
+                    dtype=np.float32,
                 ),
             }
         )
